Make it possible to use the Sentry reporter for nc14 and nc15
